requirements.txt
Django==4.1.7
django-recaptcha==3.0.0